Added database persistence

+++ streets-engine/src/main/java/net/blaklizt/streets/engine/engine/BankEngine.java
@@ -34,9 +34,9 @@ public class BankEngine extends MenuItem {
     @Override
     public Menu execute(UserSession userSession) {
 
-        return MenuItem.createFinalMenu(streets, "You currently have " + CommonUtilities.formatDoubleToMoney(
-					userSession.getUser().getUserAttribute().getBankBalance().doubleValue(),true)
-					+ " in your bank account.",
+        return MenuItem.createFinalMenu(streets, "You currently have " +
+			CommonUtilities.formatDoubleToMoney(userSession.getUser().getUserAttribute().getBankBalance(),true)
+			+ " in your bank account.",
             streets.getMainMenu(userSession), userSession);
     }
